For #483 - Refactor to make it better
